Create VirtualizedSelect component
