+++ src/components/common/Select.js
@@ -1,10 +1,12 @@
 import _ from 'lodash/fp'
+import { Children, useEffect, useRef } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import RSelect, { components as RSelectComponents } from 'react-select'
 import RAsyncCreatableSelect from 'react-select/async-creatable'
+import { AutoSizer, List } from 'react-virtualized'
 import { icon } from 'src/components/icons'
 import colors from 'src/libs/colors'
-import { useLabelAssert, useUniqueId } from 'src/libs/react-utils'
+import { useLabelAssert, useOnMount, useUniqueId } from 'src/libs/react-utils'
 
 
 const commonSelectProps = {
@@ -129,4 +131,66 @@ export const AsyncCreatableSelect = props => {
     ...commonSelectProps,
     ...props
   })
+}
+
+const VirtualizedMenuList = props => {
+  const { children, focusedOption, getStyles, getValue, maxHeight } = props
+
+  const list = useRef()
+
+  // Scroll to the currently selected value (if there is one) when the menu is opened.
+  useOnMount(() => {
+    const [value] = getValue()
+    if (value) {
+      const renderedOptions = Children.map(children, child => child.props.data)
+      const valueIndex = renderedOptions.findIndex(opt => opt.value === value.value)
+      if (valueIndex !== -1) {
+        list.current.scrollToRow(valueIndex)
+      }
+    }
+  })
+
+  // When navigating option with arrow keys, scroll the virtualized list so that the focused option is visible.
+  useEffect(() => {
+    const renderedOptions = Children.map(children, child => child.props.data)
+    const focusedOptionIndex = renderedOptions.indexOf(focusedOption)
+    if (focusedOptionIndex !== -1) {
+      list.current.scrollToRow(focusedOptionIndex)
+    }
+  }, [children, focusedOption])
+
+  const hasRenderedOptions = Array.isArray(children)
+
+  // If no options are rendered, then render the "no options" message.
+  if (!hasRenderedOptions) {
+    return children
+  }
+
+  const rowCount = children.length
+  const rowHeight = 40
+  const height = _.clamp(rowHeight, maxHeight, rowHeight * rowCount)
+
+  return h(AutoSizer, { disableHeight: true }, [
+    ({ width }) => {
+      return h(List, {
+        ref: list,
+        height,
+        width,
+        rowCount,
+        rowHeight,
+        rowRenderer: ({ index, style, key }) => div({ key, style }, [children[index]]),
+        style: { ...getStyles('menuList', props), boxSizing: 'content-box' },
+      })
+    }
+  ])
+}
+
+export const VirtualizedSelect = props => {
+  return h(Select, _.merge({
+    // react-select performs poorly when it has to render hundreds or thousands of options.
+    // This can happen for example in a workspace menu if the user has many workspaces.
+    // To address that, this component virtualizes the list of options.
+    // See https://broadworkbench.atlassian.net/browse/SUP-808 for more information.
+    components: { MenuList: VirtualizedMenuList },
+  }, props))
 }
